Consolidate UIAM conversion error codes into shared package
Moves NON_CLOUD_USER_API_KEY_CREATOR_ERROR_CODE,
API_KEY_CREATOR_NOT_ORG_MEMBER_ERROR_CODE and the
PERMANENT_UIAM_CONVERSION_ERROR_CODES list from per-plugin constants
into the shared @kbn/uiam-api-keys-provisioning-status package, so
alerting and task_manager consume them from a single source of truth.

Addresses review feedback on #271929.
