+++ x-pack/platform/packages/shared/kbn-uiam-api-keys-provisioning-status/index.ts
@@ -9,3 +9,9 @@ export {
   UiamApiKeyProvisioningEntityType,
   UiamApiKeyProvisioningStatus,
 } from './src/uiam_api_key_provisioning_status';
+
+export {
+  API_KEY_CREATOR_NOT_ORG_MEMBER_ERROR_CODE,
+  NON_CLOUD_USER_API_KEY_CREATOR_ERROR_CODE,
+  PERMANENT_UIAM_CONVERSION_ERROR_CODES,
+} from './src/uiam_conversion_error_codes';
