+++ x-pack/platform/plugins/shared/alerting/server/provisioning/constants.ts
@@ -18,20 +18,6 @@ export const TAGS = ['serverless', 'alerting', 'uiam-api-key-provisioning', 'bac
 
 export const GET_RULES_BATCH_SIZE = 300;
 export const GET_STATUS_BATCH_SIZE = 500;
-
-/**
- * UIAM convert error codes that indicate a permanent failure: rules with any of
- * these codes persisted on their `uiam_api_keys_provisioning_status` doc are
- * excluded from future provisioning attempts.
- *
- * Source: https://github.com/elastic/uiam/blob/main/modules/domain/src/main/java/co/elastic/cloud/uiam/domain/errors/ErrorCode.java
- */
-export const NON_CLOUD_USER_API_KEY_CREATOR_ERROR_CODE = '0x357391';
-export const API_KEY_CREATOR_NOT_ORG_MEMBER_ERROR_CODE = '0xBE2B58';
-export const PERMANENT_UIAM_CONVERSION_ERROR_CODES: readonly string[] = [
-  NON_CLOUD_USER_API_KEY_CREATOR_ERROR_CODE,
-  API_KEY_CREATOR_NOT_ORG_MEMBER_ERROR_CODE,
-];
 /**
  * Max number of rule IDs per KQL `or` clause when building the exclude filter.
  * Keeps each bool.should below Elasticsearch's indices.query.bool.max_clause_count (default 4096).
