+++ x-pack/platform/plugins/shared/task_manager/server/uiam_api_key_provisioning/constants.ts
@@ -31,18 +31,4 @@ export const FETCH_BATCH_SIZE = 500;
  */
 export const GET_STATUS_BATCH_SIZE = 500;
 
-/**
- * UIAM convert error codes that indicate a permanent failure: tasks with any of
- * these codes persisted on their `uiam_api_keys_provisioning_status` doc are
- * excluded from future provisioning attempts.
- *
- * Source: https://github.com/elastic/uiam/blob/main/modules/domain/src/main/java/co/elastic/cloud/uiam/domain/errors/ErrorCode.java
- */
-export const NON_CLOUD_USER_API_KEY_CREATOR_ERROR_CODE = '0x357391';
-export const API_KEY_CREATOR_NOT_ORG_MEMBER_ERROR_CODE = '0xBE2B58';
-export const PERMANENT_UIAM_CONVERSION_ERROR_CODES: readonly string[] = [
-  NON_CLOUD_USER_API_KEY_CREATOR_ERROR_CODE,
-  API_KEY_CREATOR_NOT_ORG_MEMBER_ERROR_CODE,
-];
-
 export const TAGS = ['serverless', 'task-manager', 'uiam-api-key-provisioning', 'background-task'];
